fix : Refactor the Variable Table Code, and Adjust some Table Data for another Refactor
