+++ my-app/src/components/var-table-body.module.css
@@ -1,0 +1,32 @@
+.tbody-name {
+    width: 100%;
+    height: 100%;
+    background-color: var(--sub-color);
+    border-radius: 10px;
+}
+
+.tbody-name td:hover {
+    background-color: var(--opposite-color);
+    color: var(--main-color);
+
+
+    /* box-shadow: 0px 0px 21px -1px var(--opposite-color);
+    -webkit-box-shadow: 0px 0px 21px -1px var(--opposite-color);
+    -moz-box-shadow: 0px 0px 21px -1px var(--opposite-color); */
+}
+
+/* .variable-table td:hover {
+    background-color: var(--sub-color);
+    color: var(--opposite-color);
+    box-shadow: 0px 0px 21px -1px rgba(255, 255, 255, 0.75);
+    -webkit-box-shadow: 0px 0px 21px -1px rgba(255, 255, 255, 0.75);
+    -moz-box-shadow: 0px 0px 21px -1px rgba(255, 255, 255, 0.75);
+} */
+
+.tbody-name tr:hover {
+    background-color: var(--opposite-sub-color);
+    color: var(--opposite-color);
+    /* box-shadow: 0px 0px 21px -1px var(--shadow-color);
+    -webkit-box-shadow: 0px 0px 21px -1px var(--shadow-color);
+    -moz-box-shadow: 0px 0px 21px -1px rvar(--shadow-color); */
+}
